+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- publish CSV of publications to be downloadable on MEx
+
 ### Changes
 
 ### Deprecated
@@ -33,7 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
-- use publishing endpoint for datenkompass and publisher
 - publisher is not hardcoded anymore to s3
 - added slapd healthcheck in compose file
 - update to mex-model 5.1
@@ -76,6 +77,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.18.0] - 2026-08-03
 
+### Added
+
+- link igs to seq-repo
+
 ### Fixed
 
 - use explicit paths for availability checks in datscha-web and blueant connectors
@@ -105,6 +110,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - BREAKING: Setting is renamed to ExtractorsSettings, new env prefix is MEX_EXTRACTORS_
+- updated synopse models according to mapping + CSV and renamed fields in snake_case
 
 ### Fixed
 
@@ -138,7 +144,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - implement AssetCheck `less_than_x_outbound`
 - implement AssetCheck `not_exactly_x_items`
 - implement AssetCheck `x_percent_more_than`
-- link igs to seq-repo
 
 ### Changes
 
@@ -176,7 +181,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update mex-common to 1.19.0
 - update lock file (2026_04_23)
 - seq-repo mapping update
-- updated synopse models according to mapping + CSV and renamed fields in snake_case
 
 ### Fixed